# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API for crypto-knight trading (tracks, entries, balance, orders, klines).",
        "title": "Crypto Knight API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/auth/validate": {
            "get": {
                "tags": [
                    "auth"
                ],
                "summary": "Validate auth token",
                "parameters": [
                    {
                        "type": "string",
                        "description": "JWT token",
                        "name": "token",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_antongoncharik_crypto-knight-api_internal_api_http_response.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/github_com_antongoncharik_crypto-knight-api_internal_api_http_response.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/common/off": {
            "get": {
                "tags": [
                    "common"
                ],
                "summary": "Disable bot",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_antongoncharik_crypto-knight-api_internal_api_http_response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/common/on": {
            "get": {
                "tags": [
                    "common"
                ],
                "summary": "Enable bot",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_antongoncharik_crypto-knight-api_internal_api_http_response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/common/status": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "common"
                ],
                "summary": "Get bot status",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_api_http_handler_common.StatusResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_antongoncharik_crypto-knight-api_internal_api_http_response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/entries": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "entries"
                ],
                "summary": "List all entries",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_antongoncharik_crypto-knight-api_internal_entity_entry.Entry"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_antongoncharik_crypto-knight-api_internal_api_http_response.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "entries"
                ],
                "summary": "Create entry",
                "parameters": [
                    {
                        "description": "Entry",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_antongoncharik_crypto-knight-api_internal_entity_entry.Entry"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_antongoncharik_crypto-knight-api_internal_api_http_response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_antongoncharik_crypto-knight-api_internal_api_http_response.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "github_com_antongoncharik_crypto-knight-api_internal_api_http_response.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {}
            }
        },
...
        "github_com_antongoncharik_crypto-knight-api_internal_entity_entry.Entry": {
            "type": "object",
            "required": [
                "symbol"
            ],
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "high_price": {
                    "type": "number"
                },
                "high_prices": {
                    "type": "array",
                    "items": {
                        "type": "number"
                    }
                },
                "low_price": {
                    "type": "number"
                },
                "low_prices": {
                    "type": "array",
                    "items": {
                        "type": "number"
                    }
                },
                "symbol": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "internal_api_http_handler_common.StatusResponse": {
            "type": "object",
            "properties": {
                "enabled": {
                    "type": "boolean"
                }
            }
        }
    }
}